settingPage v1.1
