+++ app/src/main/res/layout/activity_settin_page.xml
@@ -6,8 +6,10 @@
     android:layout_height="match_parent"
     tools:context=".settinPage">
 
+<!--    need to change the ImageView id-->
+
     <ImageView
-        android:id="@+id/userImage"
+        android:id="@+id/dogImage"
         android:layout_width="200dp"
         android:layout_height="200dp"
         android:layout_centerHorizontal="true"
@@ -18,7 +20,7 @@
         android:id="@+id/userID"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_below="@id/userImage"
+        android:layout_below="@id/dogImage"
         android:layout_centerHorizontal="true"
         android:layout_marginTop="30dp"
         android:text="User ID"
